HiFi-NVS/requirements_rgnv.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==0.4.3
altair==4.2.2
antlr4-python3-runtime==4.8
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
blinker==1.5
braceexpand==0.1.7
cachetools==5.3.0
carvekit-colab==4.1.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
contourpy==1.0.7
cycler==0.11.0
datasets==2.4.0
decorator==5.1.1
diffusers==0.26.3
dill==0.3.5.1
einops==0.3.0
entrypoints==0.4
executing==1.2.0
fastapi==0.95.0
fastcore==1.5.28
ffmpy==0.3.0
filelock==3.10.7
fire==0.4.0
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2024.2.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
gradio==3.21.0
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.21.4
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.0
lazy_loader==0.2
linkify-it-py==2.0.0
loguru==0.6.0
lovely-numpy==0.2.8
lovely-tensors==0.1.14
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.13
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.5.5.64
opencv-python-headless==4.7.0.72
orjson==3.8.9
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.13.1
prompt-toolkit==3.0.38
protobuf==3.20.3
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pydeck==0.8.0
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyre-extensions==0.0.23
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.4.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rich==13.3.2
rsa==4.9
safetensors==0.4.2
scikit-image==0.20.0
scipy==1.9.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
streamlit==1.20.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
test-tube==0.7.5
tifffile==2023.3.21
timm==0.6.13
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uc-micro-py==1.0.1
urllib3==1.26.15
urwid==2.1.2
uvicorn==0.21.1
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.5
websockets==10.4
Werkzeug==2.2.3
xformers==0.0.16
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
